requirements.txt
avalanche-lib==0.4.0
hydra-core==1.3.2
jsonlines==4.0.0
kornia==0.7.0
ray[default]==2.8.1
hyperopt
